chore(refactor): reduce cognitive complexity from code climate report
